🔥 refactor(sync): delete hand-written handlers package
Relocate applyEventAtomically into the domains/ package (next to its
sole consumer, ComposedSyncDomainHandler) and delete the now-empty
handlers/ package. Drop the redundant same-package import; refresh the
KDoc to name ComposedSyncDomainHandler rather than the long-gone
hand-written handlers.
