+++ sharedLogic/src/commonMain/kotlin/com/calypsan/listenup/client/data/sync/domains/ComposedSyncDomainHandler.kt
@@ -6,7 +6,6 @@ import com.calypsan.listenup.client.data.local.db.TransactionRunner
 import com.calypsan.listenup.client.data.sync.AccessFilteredSyncHandler
 import com.calypsan.listenup.client.data.sync.ClientSyncDomainRegistry
 import com.calypsan.listenup.client.data.sync.SyncDomainHandler
-import com.calypsan.listenup.client.data.sync.handlers.applyEventAtomically
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.serialization.KSerializer
